rename package directory
