--- cfggme/__init__.py
@@ -1,2 +1,0 @@
-from cfggme.compare import Comparator
-from cfggme.kilt import KILT
