+++ curvature_filtrations/compare.py
@@ -1,8 +1,11 @@
 import numpy as np
 import networkx as nx
-from cfggme.kilt import KILT
-from cfggme.topology.distances import TopologicalDistance, supported_distances
-from cfggme.topology.ph import GraphHomology
+from curvature_filtrations.kilt import KILT
+from curvature_filtrations.topology.distances import (
+    TopologicalDistance,
+    supported_distances,
+)
+from curvature_filtrations.topology.ph import GraphHomology
 from typing import List, Tuple, Optional, Union, Dict
